# api/openapi/q2-api.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "q2 API",
    "description": "Kudos (q2) — shared self-care goals. Reference API for the initial version.",
    "version": "v1"
  },
  "paths": {
...
    "/api/people/{personId}": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Returns somebody else's profile, with a balance covering only the goals you share.",
        "operationId": "GetPersonProfile",
        "parameters": [
          {
            "name": "personId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonProfileResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/images": {
      "post": {
        "tags": [
          "Images"
        ],
        "summary": "Stores an image. The body is the raw file; its media type must be image/jpeg or image/png.",
        "operationId": "UploadImage",
        "parameters": [
          {
            "name": "purpose",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ImagePurpose"
            }
          }
        ],
        "requestBody": {
          "content": {
            "image/jpeg": {
              "schema": {
                "type": "string",
                "format": "byte"
              }
            },
            "image/png": {
              "schema": {
                "type": "string",
                "format": "byte"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/images/{id}": {
      "get": {
        "tags": [
          "Images"
        ],
        "summary": "Returns the bytes of an image you are allowed to see.",
        "operationId": "GetImage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "image/jpeg": {
                "schema": {
                  "$ref": "#/components/schemas/IResult"
                }
              },
              "image/png": {
                "schema": {
                  "$ref": "#/components/schemas/IResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Images"
        ],
        "summary": "Deletes one of your images.",
        "operationId": "DeleteImage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/blocks/{personId}": {
      "post": {
        "tags": [
          "Blocks"
        ],
        "summary": "Blocks somebody and ends whatever connection there was. Blocking twice changes nothing.",
        "operationId": "BlockPerson",
        "parameters": [
          {
            "name": "personId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonSummary"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Blocks"
        ],
        "summary": "Lifts a block you set. The friendship does not come back.",
        "operationId": "UnblockPerson",
        "parameters": [
          {
            "name": "personId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonSummary"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notifications/subscribe": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Registers this browser, or brings its keys up to date.",
        "operationId": "SubscribeToNotifications",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/unsubscribe": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Forgets this browser. Succeeds whether or not there was anything to forget.",
        "operationId": "UnsubscribeFromNotifications",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnsubscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BalanceResponse": {
        "required": [
          "done",
          "missed"
        ],
        "type": "object",
        "properties": {
          "done": {
            "type": "integer",
            "format": "int32"
          },
          "missed": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "FriendshipState": {
        "enum": [
          "None",
          "RequestSent",
          "RequestReceived",
          "Friends",
          "Self"
        ]
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "IResult": {
        "type": "object"
      },
      "ImagePurpose": {
        "enum": [
          "Avatar",
          "Proof",
          "ChallengeEntry"
        ]
      },
...
      "ImageResponse": {
        "required": [
          "id",
          "purpose",
          "width",
          "height",
          "byteSize",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "purpose": {
            "$ref": "#/components/schemas/ImagePurpose"
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "byteSize": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PersonProfileResponse": {
        "required": [
          "person",
          "state",
          "streak",
          "kudosReceived",
          "goalsCompleted",
          "balance",
          "sharedGoals"
        ],
        "type": "object",
        "properties": {
          "person": {
            "$ref": "#/components/schemas/PersonSummary"
          },
          "state": {
            "$ref": "#/components/schemas/FriendshipState"
          },
          "streak": {
            "type": "integer",
            "format": "int32"
          },
          "kudosReceived": {
            "type": "integer",
            "format": "int32"
          },
          "goalsCompleted": {
            "type": "integer",
            "format": "int32"
          },
          "balance": {
            "$ref": "#/components/schemas/BalanceResponse"
          },
          "sharedGoals": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PersonSummary": {
        "required": [
          "id",
          "displayName",
          "handle",
          "initials",
          "avatarColor",
          "isOnline",
          "avatarImageId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "handle": {
            "type": "string"
          },
          "initials": {
            "type": "string"
          },
          "avatarColor": {
            "type": "string"
          },
          "isOnline": {
            "type": "boolean"
          },
          "avatarImageId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "SubscribeRequest": {
        "type": "object",
        "properties": {
          "endpoint": {
            "type": "string",
            "nullable": true
          },
          "publicKey": {
            "type": "string",
            "nullable": true
          },
          "authSecret": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "UnsubscribeRequest": {
        "type": "object",
        "properties": {
          "endpoint": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  },
...
}